Switch to faster update mode when file is above configured threshold
Unfortunately, using the 'undoable writer' to update the edit buffer via
the ToolsAPI is very slow, tending towards unusable on large files.

In my testing I found that the lag becomes noticable for files above
1MiB, and the editor becomes completely unresponsive for files around
10MiB.

There is a faster method for updating the buffer, but it destroys the
undo history. On balance, losing the undo history seems like a fair
compromise for keeping the editor responsive. However, it's not possible
for us to say at what file size these performance issues become a
problem on other machines and other files, so I've created a setting to
control the file size threshold.